Moved folders to fileParsing under feature and limit test scope to it
