+++ src/backend/features/fileParsing/errors/traverserError.ts
