+++ src/backend/features/fileParsing/tests/fileParsingService.test.ts
@@ -1,9 +1,11 @@
 import { describe, expect, it } from "vitest";
 
 import { TraverserError } from "../errors/traverserError";
-import { getExportsFromFilesInFolder } from "../fileParsingService";
+import { getExportsFromFilesInFolder } from "../services/fileParsingService";
+
 
 const testFolder = "testResources";
+const featureFolderName = "fileParsing"
 
 /**
 	 * Expected folder structure:
@@ -16,7 +18,7 @@ const testFolder = "testResources";
 
 describe("getExportsFromFilesInFolder", () => {
 	it("Given above folder structure should find at least one export", async () => {
-		const exports = await getExportsFromFilesInFolder(testFolder);
+		const exports = await getExportsFromFilesInFolder(testFolder,{ascendToThisFolderBeforeStartingSearch: featureFolderName});
 		expect(Array.isArray(exports)).toBe(true);
 		expect(exports.length).not.toBe(0);
 		exports.forEach((exportObj) => {
@@ -27,6 +29,7 @@ describe("getExportsFromFilesInFolder", () => {
 	it("Given above folder structure should find at least 2 exports", async () => {
 		const exports = await getExportsFromFilesInFolder(testFolder, {
 			fileFilter: (filePath) => filePath.endsWith(".ts"),
+			ascendToThisFolderBeforeStartingSearch: featureFolderName
 		});
 		expect(Array.isArray(exports)).toBe(true);
 		expect(exports.length).toBeGreaterThan(1)
@@ -36,18 +39,19 @@ describe("getExportsFromFilesInFolder", () => {
 		
 		const exports = await getExportsFromFilesInFolder(testFolder, {
 			fileFilter: (filePath) => filePath.endsWith("testfile2.ts"),
+			ascendToThisFolderBeforeStartingSearch: featureFolderName
 		});
 		expect(Array.isArray(exports)).toBe(true);
 		expect(exports).toHaveLength(1);
 	});
 
 	it("should throw an error if the specified folder is not found", () => {
-		expect(() => getExportsFromFilesInFolder("nonExistentFolder"))
+		expect(() => getExportsFromFilesInFolder("nonExistentFolder",{ascendToThisFolderBeforeStartingSearch: featureFolderName}))
 			.toThrow(TraverserError);
 	});
 
 	it("Given above folder structure, should be able to locate the contents of testfile3.ts with recursive search", async () => {
-		const exports = await getExportsFromFilesInFolder(testFolder,{recursive:true});
+		const exports = await getExportsFromFilesInFolder(testFolder,{recursive:true,ascendToThisFolderBeforeStartingSearch: featureFolderName});
 		expect(Array.isArray(exports)).toBe(true);
 		expect(exports.length).not.toBe(0);
 		const res = exports.filter(obj => obj.testVar3 === "this is test file 3")
@@ -55,7 +59,7 @@ describe("getExportsFromFilesInFolder", () => {
 	});
 
 	it("Given above folder structure should not be able to find the contents of testfile3.ts without recursion", async () => {
-		const exports = await getExportsFromFilesInFolder(testFolder);
+		const exports = await getExportsFromFilesInFolder(testFolder,{ascendToThisFolderBeforeStartingSearch: featureFolderName});
 		expect(Array.isArray(exports)).toBe(true);
 		expect(exports.length).not.toBe(0);
 		const res = exports.filter(obj => obj.testVar3 === "this is test file 3")
